sdk/tests/conformance/manual/framebuffers-keep-contents-exiting-fullscreen-mode.html

<!--

/*
** Copyright (c) 2013 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Check that framebuffers keep contents exiting fullscreen mode.</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
<style>
canvas {
    width: 10%;
    height: 10%;
    border: 1px solid black;
}
body,html {
    width: 100%;
    height: 100%;
    padding: 0px;
    margin: 0px;
    border: 0px;
}
#f {
    z-index: 2;
    position: absolute;
    left: 0px;
    top: 0px;
    width: 100%;
    height: 100%;
    text-align: center;
}
#inner {
    position: relative;
    top: 50%;
}
#button {
    background-color: white;
    border: 1px solid black;
    font-size: 24pt;
    padding: 1em;
    margin-left: auto;
    margin-right: auto;
}
</style>
</head>
<body>
<pre>
This test must be run manually.

Checks that framebuffers keep their contents going into and out of fullscreen mode.

Click the button.
</pre>
<div id="f">
  <div id="inner">
    <div id="buttonHolder">
      <div id="button">click to <span id="mode">enter</span> fullscreen</div>
      <canvas id='c'></canvas>
    </div>
  </div>
</div>
<div id="console"></div>
<script>
"use strict";
var wtu = WebGLTestUtils;
var testedFullScreen = false;
var mode = document.getElementById("mode");
var c = document.getElementById("c");
var button = document.getElementById("button");
var buttonHolder = document.getElementById("buttonHolder");
var gl = wtu.create3DContext(c);
if (!gl) {
  testFailed("can't init WebGL");
}

var checkState = function() {
  debug("");
  wtu.checkCanvasRect(gl, 0, 0, 1, 1, [0, 255, 0, 255]);
  shouldBeNonNull("gl.getParameter(gl.FRAMEBUFFER_BINDING)");
  glErrorShouldBe(gl, gl.NO_ERROR, "there should be no errors");
}

var checkFramebufferState = function(fullscreen) {
  checkState();

  debug("fullscreen:" + fullscreen);
  mode.innerText = fullscreen ? "exit" : "enter";

  if (fullscreen) {
    // Not sure if this is needed but need to make sure
    // we at least went fullscreen once.
    testedFullScreen = true;
  } else {
    if (testedFullScreen) {
      buttonHolder.innerHTML = "";
      finishTest();
    }
  }
};

var test = function() {
  if (!wtu.setupFullscreen("button", "f", checkFramebufferState)) {
    testPassed("Browser does not support fullscreen mode. This is OK");
    finishTest();
    return;
  }

  var fb = gl.createFramebuffer();
  var tex = gl.createTexture();

  gl.clearColor(1,0,0,1);
  gl.clear(gl.COLOR_BUFFER_BIT);

  gl.bindFramebuffer(gl.FRAMEBUFFER, fb);
  gl.bindTexture(gl.TEXTURE_2D, tex);
  gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, 1, 1, 0, gl.RGBA, gl.UNSIGNED_BYTE, null);
  gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex, 0);
  shouldBe("gl.checkFramebufferStatus(gl.FRAMEBUFFER)", "gl.FRAMEBUFFER_COMPLETE");

  gl.clearColor(0, 1, 0, 1);
  gl.clear(gl.COLOR_BUFFER_BIT);

  checkState();
};
test();

</script>
</body>
</html>
